chore: isolate the workflow overview  from execution page
